cleanup UX layout code. fix alignment for U6
