+++ com.unity.cinemachine/Editor/Editors/CinemachineStoryboardEditor.cs
@@ -24,7 +24,7 @@ namespace Unity.Cinemachine.Editor
                 tooltip = CinemachineCorePrefs.s_StoryboardGlobalMuteLabel.tooltip,
                 value = CinemachineStoryboard.s_StoryboardGlobalMute
             });
-            muteToggle.AddToClassList(InspectorUtility.kAlignFieldClass);
+            muteToggle.AddToClassList(InspectorUtility.AlignFieldClassName);
             muteToggle.RegisterValueChangedCallback((evt) => 
             {
                 CinemachineStoryboard.s_StoryboardGlobalMute = CinemachineCorePrefs.StoryboardGlobalMute.Value = evt.newValue;
@@ -35,7 +35,7 @@ namespace Unity.Cinemachine.Editor
             var showImageProp = serializedObject.FindProperty(() => Target.ShowImage);
             var row = ux.AddChild(new InspectorUtility.LabeledRow("Show Image", showImageProp.tooltip));
             var imageToggle = row.Contents.AddChild(new Toggle("") 
-                { style = { flexGrow = 0, marginTop = 3, marginLeft = 0, alignSelf = Align.Center }});
+                { style = { flexGrow = 0, marginTop = 3, marginLeft = 6, alignSelf = Align.Center }});
             imageToggle.BindProperty(showImageProp);
             var imageField = row.Contents.AddChild(new PropertyField(serializedObject.FindProperty(() => Target.Image), "")
                 { style = { flexGrow = 1, marginLeft = 4 }});
@@ -49,11 +49,10 @@ namespace Unity.Cinemachine.Editor
             // Scale
             var scaleProp = serializedObject.FindProperty(() => Target.Scale);
             row = ux.AddChild(new InspectorUtility.LabeledRow(scaleProp.displayName, scaleProp.tooltip));
-            var scaleField = row.Contents.AddChild(new PropertyField(scaleProp, "") 
-                { style = { flexBasis = 0, flexGrow = 2, marginLeft = -3 }});
+            var scaleField = row.Contents.AddChild(new PropertyField(scaleProp, "") { style = { flexBasis = 50, flexGrow = 2 }});
             var syncProp = serializedObject.FindProperty(() => Target.SyncScale);
             var syncField = row.Contents.AddChild(new Toggle("") 
-                { text = " Sync", style = { flexBasis = 0, flexGrow = 1, paddingLeft = 3, paddingTop = 3 }});
+                { text = " Sync", style = { flexBasis = 20, flexGrow = 1, paddingLeft = 1, paddingTop = 2 }});
             syncField.BindProperty(syncProp);
             scaleField.TrackPropertyWithInitialCallback(syncProp, (p) =>
             {
