+++ com.unity.cinemachine/Editor/Obsolete/ObsoleteInspectorUtility.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e431aac1b3a84b64e93b38e833c448be
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
