+++ com.unity.cinemachine/Editor/PropertyDrawers/OutputChannelsPropertyDrawer.cs
@@ -20,7 +20,7 @@ namespace Unity.Cinemachine.Editor
 
             var row = new InspectorUtility.LabeledRow(preferredLabel, property.tooltip);
             var selector = row.Contents.AddChild(new MaskField(choices, 1) 
-                { tooltip = property.tooltip, style = { flexBasis = 100, flexGrow = 1, marginLeft = 0 }});
+                { tooltip = property.tooltip, style = { flexBasis = 100, flexGrow = 1, marginLeft = 3 }});
             selector.BindProperty(property);
 
             row.Contents.Add(InspectorUtility.MiniPopupButton(null, new ContextualMenuManipulator((evt) =>
