+++ com.unity.cinemachine/Editor/PropertyDrawers/SensorSizePropertyDrawer.cs
@@ -75,8 +75,7 @@ namespace Unity.Cinemachine.Editor
                 property.serializedObject.ApplyModifiedProperties();
 
                 // Find matching preset (if any)
-                var index = s_PresetSizes.FindIndex(
-                    e => Mathf.Approximately(v.x, e.x) && Mathf.Approximately(v.y, e.y));
+                var index = s_PresetSizes.FindIndex(e => Mathf.Approximately(v.x, e.x) && Mathf.Approximately(v.y, e.y));
                 presets.SetValueWithoutNotify(index >= 0 ? s_PresetNames[index] : "Custom");
             });
