+++ com.unity.cinemachine/Runtime/Behaviours/CinemachineTargetGroup.cs
@@ -173,7 +173,7 @@ namespace Unity.Cinemachine
         //============================================
         // Legacy support 
 
-        [SerializeField, FormerlySerializedAs("m_Targets")]
+        [HideInInspector, SerializeField, FormerlySerializedAs("m_Targets")]
         Target[] m_LegacyTargets;
 
         void Awake()
